Feat/ATM-87 USDT on XTZ (#111)
* Add USDT_XTZ Fa2 token

* Temporary add Fa2 swap logic based on Fa12 swap logic

* Add USDT_XTZ currency and symbols for testnet

* Change USDT descriptions

* Fa2Account IEstimatable implementation

* Мoving the general configuration for tezos tokens to TezosTokenConfig

* Unify Fa12 and Fa2 accounts

* Fix AddressesHelper

* Fix Fa12 testnet configuration

* Fa2Swap helpers

* Add FA2 operator check method

* Fix IsFa2TokenOperatorActiveAsync

* Fix Tezos tokens balance checker, fix update_operators arguments

* Increase gas limit for testnet USDT

* Fix pending status for Ethereum transactions

* Make the DefaultDataFileName constant as public in the Account class

* Fix WebSocketAtomexClientLegacy contructor

* Fix USDt displayed name

* Fix swaps logic for compatibility with new and legacy Atomex backend

* Add acceptor requisites check and save

* Fix purchased currency displayed name is swap details

* Update currencies config to v8

+++ Atomex.Client.Core/Wallet/Abstract/IAccount.cs
@@ -83,14 +83,14 @@ namespace Atomex.Wallet.Abstract
         ICurrencyAccount GetTezosTokenAccount(
             string currency,
             string tokenContract,
-            decimal tokenId);
+            int tokenId);
 
         T GetCurrencyAccount<T>(string currency) where T : class, ICurrencyAccount;
 
         T GetTezosTokenAccount<T>(
             string currency,
             string tokenContract,
-            decimal tokenId) where T : class;
+            int tokenId) where T : class;
 
         string GetUserId(uint keyIndex = 0);
